Created interface for CatalogService to allow for multiple implementations. Fixes #28.
